+++ cs/TagsCloudVisualization/PointFExtensions.cs
@@ -11,5 +11,12 @@ namespace TagsCloudVisualization
             var dy = to.Y - from.Y;
             return (float)Math.Sqrt(dx * dx + dy * dy);
         }
+
+        public static RectangleF GetRectangle(this PointF center, Size size)
+        {
+            return new RectangleF(
+                new PointF(center.X - size.Width / 2f, center.Y - size.Height / 2f),
+                size);
+        }
     }
 }
