+++ cs/TagsCloudVisualization/Word/WordsHandler.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Drawing;
+using TagsCloudVisualization.CloudLayouter;
 
 namespace TagsCloudVisualization
 {
